+++ src/main/java/com/kometsales/inventory/entity/BoxTypeEntity.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -20,15 +19,15 @@ public class BoxTypeEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    private Integer id;
+    private int id;
     @Column(name = "code", length = 80)
     private String code;
     @Column(name = "width", nullable = false, precision = 10, scale = 5)
-    private BigDecimal width;
+    private double width;
     @Column(name = "height", nullable = false, precision = 10, scale = 5)
-    private BigDecimal height;
+    private double height;
     @Column(name = "length", nullable = false, precision = 10, scale = 5)
-    private BigDecimal length;
+    private double length;
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "boxTypeEntity")
     @JsonIgnore
     private Set<InventoryEntity> inventoryEntities = new HashSet<>(0);
